requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
anyio==4.12.1
bandit==1.8.6
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.19.1
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.3
macholib @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pluggy==1.6.0
Pygments==2.19.2
pypdf==6.9.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/39d9dc1a-2111-11f0-be06-226177e5bb69/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
stevedore==5.5.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.1
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vboxapi==1.0
